gestion commande commencé, et gestion déchets commencé aussi

+++ Maison_moel/Entities/Dechet.cs
@@ -7,7 +7,7 @@ public partial class Dechet
 {
     public int IdDechets { get; set; }
 
-    public int Poids { get; set; }
+    public float Poids { get; set; }
 
     public DateTime DateSaisie { get; set; }
 }
